requirements.txt
astroid==2.2.5
Babel==2.7.0
boto3==1.9.167
botocore==1.12.167
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
django-phone-field==1.7.1
django-phonenumber-field==3.0.1
docutils==0.14
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
isort==4.3.20
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
Phon==0.1a3
phonenumbers==8.10.13
psycopg2==2.8.3
psycopg2-binary==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.11.1
